Release via GH Actions CI, drop Drone support

+++ .github/workflows/cleanup.sh
@@ -6,8 +6,10 @@ for folder in $TARGETS; do
   for f in $(ls -a); do
     if [ "$f" != "." ] &&
        [ "$f" != ".." ] &&
-       [ "$f" != "_github_home" ]  # __w/_temp/_github_home is mount as
-                                   # /github/home and is cleaned separately
+       [ "$f" != "_github_home" ] && # __w/_temp/_github_home is mount as
+                                     # /github/home and is cleaned separately
+       [ "$f" != "_github_workflow" ]  # Contains workflow metadata, is needed
+                                       # for the correct workflow execution
     then
       rm -rf "$f"
     fi
